+++ src/Seomatic.php
@@ -63,6 +63,7 @@ use nystudio107\seomatic\helpers\Environment;
 use nystudio107\seomatic\helpers\Environment as EnvironmentHelper;
 use nystudio107\seomatic\helpers\Gql as GqlHelper;
 use nystudio107\seomatic\helpers\MetaValue as MetaValueHelper;
+use nystudio107\seomatic\helpers\Schema as SchemaHelper;
 use nystudio107\seomatic\integrations\feedme\SeoSettings as SeoSettingsFeedMe;
 use nystudio107\seomatic\models\MetaScriptContainer;
 use nystudio107\seomatic\models\Settings;
@@ -302,6 +303,7 @@ class Seomatic extends Plugin
         self::$plugin->frontendTemplates->invalidateCaches();
         self::$plugin->metaContainers->invalidateCaches();
         self::$plugin->sitemaps->invalidateCaches();
+        SchemaHelper::invalidateCaches();
         // Clear the GraphQL caches too
         $gql = Craft::$app->getGql();
         if (method_exists($gql, 'invalidateCaches')) {
@@ -1071,6 +1073,12 @@ class Seomatic extends Plugin
                 'key' => 'seomatic-sitemap-caches',
                 'label' => Craft::t('seomatic', 'SEOmatic sitemap caches'),
                 'action' => [self::$plugin->sitemaps, 'invalidateCaches'],
+            ],
+            // Schema caches
+            [
+                'key' => 'seomatic-schema-caches',
+                'label' => Craft::t('seomatic', 'SEOmatic schema caches'),
+                'action' => [SchemaHelper::class, 'invalidateCaches'],
             ],
         ];
     }