+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # SEOmatic Changelog
 
+## 4.0.22 - 2023.03.13
+### Added
+* Added `(Google rich result)` in the Main Entity of Page dropdown, for Schema.org types that Google uses for Rich Results
+* Add descriptions & links for `(pending)` and `(Google rich result)` schemas in the Main Entity of Page dropdown
+* Added a caching layer to the Schema helper
+
 ## 4.0.21 - 2023.03.09
 ### Added
 * Added support for Doxster field types as pull sources ([#1279](https://github.com/nystudio107/craft-seomatic/pull/1279))